feat: 1884-EggDropWith2EggsAndNFloors
Quite a difficult DP problem in my opinion.

O(kn^2) solution where k is the number of eggs - 2 in this problem, and
n is the number of floors.

There are actually solutions which have grows asymptotically slower than
this, and requires a different parameterization/state. Will be exploring
these variants soon.