feat: 42-TrappingRainWater
First LC Hard.

There is a 2-pointer approach that I know exists for this problem,
but that approach is not as generalizable as this monotonic Stack
approach, and therefore has less learning value IMO.

Quite a tricky problem consisting of multiple parts. The key to this
problem is that, suppose we have a left boundary, we need to find a
right boundary that is equal to or greater than our left boundary, for
water to be trapped.

The calculation of the exact unit amount of water trapped can be
visualized as a histogram, where due to the monotonically decreasing
property of our stack, we can inductively add to the resultant amount of
water trapped. To elaborate, suppose we have some arbitrary indices
{0, 1, 2} representing elevations {3, 2, 1} in the Stack, and we
encounter an elevation of 4 at index 3.

We first pop the most minimum value in the stack, with (index,
elevation) corresponding to (2, 1). Then, looking to its left, we have
(1, 2). The amount of water trapped by this set of elevations is simply
the difference of min(leftBound, rightBound) = min(2, 4) = 2 and the
popped value 1. Thus we get 1 unit of water trapped.

The next iteration, we pop (1, 2) and look at (0, 3). The height of
water bounded will be
`min(leftBound, rightBound) - popped = min(3, 4) = 3 - 2 = 1 unit`
As such, we add 1 unit of water to the result. However, this does not
account for the previously popped (2, 1), which has a lower elevation
than the current popped (1, 2), and will therefore have the same height
of trapped water. Thus we find the horizontal distance between the
leftBound and rightBound, denoted by `i - len(indices) - 1`, and
multiply it by the height of water added in the current iteration,
giving us the correct result of 3 units of water trapped for the current
iteration.