feat: 881-BoatsToSavePeople
DP question that has a Greedy Choice Property.
Optimal substructure is not as intuitive as the Greedy Choice Property,
although both are a little hard translate from intuition to words.

Attempted to optimise runtime using a counting-based sorting algorithm
instead of a comparison-based sort, but benchmarks showed a contrasting
result where the comparison-based implementation ran faster than the
counting-based one.
This is due to the small test case used which really highlighted the
misuse of complexity analysis as a way to determine which
implementations would run faster, when it is really measuring the rate
of growth of the runtime with respect to the input size.

Similar variant covered in CS3230 Greedy Algorithms Tutorial.