feat(C++): Add 719
A really difficult question combining multiple concepts. Being able to
binary search the difference is absolutely not obvious at all. It
requires identifying that the range of possible distances is finite and
bounded. It requires identifying the monotonic property that the number
of pairs with distance <= d increases as the distance d increases (in
hindsight, its not that much of a logical leap).