feat(C++): Add 1937
Very interesting problem. Requires multiple insights to solve. The naive
DP solution using a nested O(n) loop to iterate over the previous row
and calculate the would-be sum + penalty involves many repeated
computation.

However, identifying that was tricky. Finding a solution to
solve that was also tricky! Involves framing the sub-problem of the
re-computation as a "Maximum points of the first jth ..." and the
"Maximum points of the last jth ...", i.e. a prefix/postfixMax.