feat(C++): Add 407
Interesting Hard question. The key insight is that border cells cannot
trap water. As such, the traversal should be from outwards to inwards.

Another key-insight is that water can drain-off. So as long as the
current considered border has any lower-height cells, water cannot be
trapped. As such, the traversal should be from lower-height cells to
higher-height cells.