feat(C++): Add 1143
1D DP implementation is quite tricky. Had some issues with
implementation, but basically has to use a temporary array to represent
the LCS of the current row. This is because the previous diagonal was
referenced, and could be overwritten.

Can be optimized further, since only 3 variables are used at any point:
dp[i-1][j-1], dp[i][j-1] and dp[i-1][j].