feat(C++): Add 1730
Interestingly, goal-checks on popping will result in TLE.
As such, it is necessary to do checks before adding to the frontier for
optimization.
This is only possible because the edges are of uniform weight; the first
time the goal is encountered, it is guaranteed to be at the shortest
distance.