feat(C++): Add 1545
There is an O(1) solution that utilises the relationship between the set
bits of K and the length of the Nth string.

Essentially, from the MSB to the LSB, a '0' bit indicates that the kth
bit is in the left half, while a '1' bit indicates that the kth bit is
in the right half, allowing elimination of half of the string.

Since there are at most 32 bits in K, this is an O(1) time/space
solution.

To revisit.