feat(dijkstra): Add 2577
Identifying that it was Dijkstra was relatively simple.
What was difficult was determining whether a search path can be pruned.
Typically, we try to avoid revisiting cells in a Graph indiscriminately.
We want to be able to use some sort of heuristics to prune.

In this case, identifying that we can "wait" to visit the next cell was
crucial, since it allowed us to reduce the heuristics to simply
visited|not visited (once popped off the PQ, it is guaranteed that the
shortest path to that node has been found).

Quite interesting.