feat(line-sweep): Add 218
Core idea was relatively straightforward, with knowledge of Line Sweep
technique.

Implementation-wise was not as smooth though. Logical errors included
the slightly unusual ordering of start events before end events, but
that was a smaller issue.

The bigger issue was std::multiset::erase removing all elements matching
the given key...
Either do std::multiset::extract which removes a single node, or the
overloaded std::multiset::erase that takes in an iterator, which
requires a predecessor call to std::multiset::find.

Overall, interesting question.